Added tests for alignment effects.
I added seperate aligned compelx datatypes instead of just
typedefing the complex types to be more aligned. Although the later
is suggested on page 196 of the OpenCL Standard version 1.1, this
will not give the same effect on AMD hardware. Actually, once a type
is typedefed you won't even see the effect if the original type has
an alignment specified.